Remove Exception and replace with logging user out.
This will only happen if a user is logged in but is not in the database due to some error.
In this case the application should handle it in a user-friendly way and log them out.
Before, it threw an exception which makes a user unable to use the program.